# teacher/backend/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: TIDE Teacher API Contract
  version: 0.10.0
  description: >-
    教师端公共合同。包含教师端任务、文件和 My TIDE 安全读取；
    各模块实现状态见 contracts/README。任务、G01 状态和外部消息直接使用共享 PostgreSQL 契约表；
    教师积分与逐课积分直接读取世文当前聚合视图，不在教师端重算。
servers:
  - url: http://localhost:3000
    description: local backend
...
paths:
...
  /api/v1/me/profile:
    get:
      tags: [MyTIDE]
      summary: 获取当前教师账号和训练营安全资料
      operationId: getTeacherProfile
      security: [{ bearerAuth: [] }]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 30 }
        - name: cursor
          in: query
          schema: { type: string, maxLength: 512 }
        - name: filter
          in: query
          schema: { type: string, enum: [ALL, UNREAD], default: ALL }
      responses:
        '200':
          description: 世文实时资料或最近一次安全缓存
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TeacherProfile' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/me/g01-review:
    get:
      tags: [MyTIDE, Tasks]
      summary: 获取 G01 Self-intro 和 TESOL 外部审核安全状态
      operationId: getG01Review
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: 后端映射后的教师安全审核状态
          content:
            application/json:
              schema: { $ref: '#/components/schemas/G01Review' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
  /api/v1/me/tide-summary:
    get:
      tags: [MyTIDE]
      summary: 读取世文当前教师积分卡
      operationId: getTideSummary
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: 直接透传世文当前总分、资格、五维及子项；可获得积分仅由未完成 G 任务生成
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TideSummary' }
  /api/v1/me/courses:
    get:
      tags: [MyTIDE]
      summary: 分页读取世文当前逐课积分明细
      operationId: getCourseFacts
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: 无历史数据时 items 为空数组
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CourseList' }
  /api/v1/me/notifications:
    get:
      tags: [Notifications]
      summary: 获取教师通知
      operationId: getNotifications
      security: [{ bearerAuth: [] }]
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 30 }
        - name: cursor
          in: query
          required: false
          schema: { type: string, maxLength: 512 }
        - name: filter
          in: query
          required: false
          schema: { type: string, enum: [ALL, UNREAD], default: ALL }
      responses:
        '200':
          description: 合并返回共享外部消息与 tide 系统消息
          content:
            application/json:
              schema:
                type: object
                required: [items, totalCount, unreadCount, nextCursor, freshness]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Notification' }
                  totalCount: { type: integer, minimum: 0 }
                  unreadCount: { type: integer, minimum: 0 }
                  nextCursor: { type: [string, 'null'] }
                  freshness: { $ref: '#/components/schemas/SourceFreshness' }
  /api/v1/me/notifications/{sourceNotificationId}/read:
    post:
      tags: [Notifications]
      summary: 幂等回写外部或系统消息已读状态
      operationId: markNotificationRead
      security: [{ bearerAuth: [] }]
      parameters:
        - name: sourceNotificationId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: 已读时间已保存 }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/me/notifications/{sourceNotificationId}/click:
    post:
      tags: [Notifications]
      summary: 幂等回写外部或系统消息点击状态
      operationId: markNotificationClicked
      security: [{ bearerAuth: [] }]
      parameters:
        - name: sourceNotificationId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: 点击时间已保存 }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    NotFound:
      description: 资源不存在或不属于当前教师
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
    ServiceUnavailable:
      description: 数据源或外部依赖暂时不可用
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
...
    TeacherProfile:
      type: object
      additionalProperties: false
      required: [teacherId, email, name, timezone, campDay, totalCampDays, graduationState, freshness]
      properties:
        teacherId: { type: string }
        email: { type: string, format: email }
        name: { type: string }
        timezone: { type: [string, 'null'] }
        campDay: { type: [integer, 'null'], minimum: 0 }
        totalCampDays: { type: integer, minimum: 1 }
        graduationState: { type: [string, 'null'] }
        freshness: { $ref: '#/components/schemas/SourceFreshness' }
    ExternalReviewStatus:
      type: string
      enum: [WAITING, IN_REVIEW, APPROVED, NEEDS_CHANGES, UNAVAILABLE]
    G01Review:
      type: object
      additionalProperties: false
      required: [selfIntroStatus, tesolStatus, externalStatusesComplete, freshness]
      properties:
        selfIntroStatus: { $ref: '#/components/schemas/ExternalReviewStatus' }
        tesolStatus: { $ref: '#/components/schemas/ExternalReviewStatus' }
        externalStatusesComplete:
          type: boolean
          description: 仅表示 Self-intro 与 TESOL 两项外部状态均完成，不代表 G01 整体完成。
        freshness: { $ref: '#/components/schemas/SourceFreshness' }
...
    SourceFreshness:
      type: object
      required: [source, sourceUpdatedAt, fetchedAt, stale]
      properties:
        source: { type: string, enum: [LIVE, CACHED, UNAVAILABLE] }
        sourceUpdatedAt: { type: [string, 'null'], format: date-time }
        fetchedAt: { type: [string, 'null'], format: date-time }
        stale: { type: boolean }
    TideSummary:
      type: object
      additionalProperties: false
      required: [available, rawTotalScore, publicTotalScore, graduationState, graduationQualified, goldQualified, graduationThreshold, goldThreshold, mandatoryTaskCompletedCount, mandatoryTaskTotalCount, scoreRuleVersion, calculatedAt, dimensions, availableScore, freshness]
      properties:
        available: { type: boolean, const: true }
        rawTotalScore: { type: number, minimum: 0 }
        publicTotalScore: { type: number, minimum: 0, maximum: 200 }
        graduationState: { type: string }
        graduationQualified: { type: boolean }
        goldQualified: { type: boolean }
        graduationThreshold: { type: number, minimum: 0 }
        goldThreshold: { type: number, minimum: 0 }
        mandatoryTaskCompletedCount:
          type: integer
          minimum: 0
          description: 当前 G01–G09 九项必修任务中已完成的数量。
        mandatoryTaskTotalCount:
          type: integer
          minimum: 0
          description: 当前固定必修任务总数，正常为 9。
        scoreRuleVersion: { type: string }
        calculatedAt: { type: string, format: date-time }
        dimensions:
          type: array
          minItems: 5
          maxItems: 5
          items: { $ref: '#/components/schemas/ScoreDimension' }
        availableScore:
          oneOf:
            - { $ref: '#/components/schemas/AvailableScore' }
            - { type: 'null' }
        freshness: { $ref: '#/components/schemas/SourceFreshness' }
    ScoreDimension:
      type: object
      additionalProperties: false
      required: [code, score, scoreRuleVersion, projectionRevision, calculatedAt, components]
      properties:
        code:
          type: string
          enum: [USER_FEEDBACK, RELIABILITY, CLASS_QUALITY, CAPACITY, NEW_TEACHER_TASK]
        score: { type: number, minimum: 0 }
        scoreRuleVersion: { type: string }
        projectionRevision: { type: integer, minimum: 0 }
        calculatedAt: { type: string, format: date-time }
        components:
          type: array
          items: { $ref: '#/components/schemas/ScoreComponent' }
    ScoreComponent:
      type: object
      additionalProperties: false
      required: [code, unitCount, pointsPerUnit, score, sourceScope, sourceMetric]
      properties:
        code: { type: string }
        unitCount: { type: number, minimum: 0 }
        pointsPerUnit: { type: [number, 'null'], minimum: 0 }
        score: { type: number, minimum: 0 }
        sourceScope: { type: string, enum: [LESSON, TEACHER, TASK] }
        sourceMetric: { type: [string, 'null'] }
    AvailableScore:
      type: object
      description: 未进入终态的 G01–G09 可获得积分；隐藏 G00 不进入。
      additionalProperties: false
      required: [score, items]
      properties:
        score: { type: number, exclusiveMinimum: 0 }
        items:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/AvailableScoreItem' }
    AvailableScoreItem:
      type: object
      additionalProperties: false
      required: [taskCode, title, score, taskStatus]
      properties:
        taskCode: { type: string }
        title: { type: [string, 'null'] }
        score: { type: number, exclusiveMinimum: 0 }
        taskStatus: { type: string }
    CourseList:
      type: object
      additionalProperties: false
      required: [items, page, pageSize, totalCount, freshness]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/CourseFact' }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 100 }
        totalCount: { type: integer, minimum: 0 }
        freshness: { $ref: '#/components/schemas/SourceFreshness' }
    CourseFact:
      type: object
      additionalProperties: false
      required: [lessonId, lessonSequence, lessonCount, scheduledStartAt, lessonLocalDate, lessonLocalTime, lifecycleStatus, validForScoring, evidenceStatus, lessonTotalScore, scoreRuleVersion, updatedAt, facts, dimensions]
      properties:
        lessonId: { type: string }
        lessonSequence: { type: integer, minimum: 1 }
        lessonCount: { type: integer, minimum: 0 }
        scheduledStartAt: { type: [string, 'null'], format: date-time }
        lessonLocalDate: { type: [string, 'null'], format: date }
        lessonLocalTime: { type: [string, 'null'] }
        lifecycleStatus: { type: string }
        validForScoring: { type: boolean }
        evidenceStatus: { type: string }
        lessonTotalScore: { type: number, minimum: 0 }
        scoreRuleVersion: { type: string }
        updatedAt: { type: string, format: date-time }
        facts:
          type: object
          additionalProperties: false
          required: [late, earlyLeave, falseEarlyLeave, positiveFeedback, favorited, rebooked, cameraOff, cpuUsageHigh, networkDelayHigh, peak]
          properties:
            late: { type: [boolean, 'null'] }
            earlyLeave: { type: [boolean, 'null'] }
            falseEarlyLeave: { type: [boolean, 'null'] }
            positiveFeedback: { type: [boolean, 'null'] }
            favorited: { type: [boolean, 'null'] }
            rebooked: { type: [boolean, 'null'] }
            cameraOff: { type: [boolean, 'null'] }
            cpuUsageHigh: { type: [boolean, 'null'] }
            networkDelayHigh: { type: [boolean, 'null'] }
            peak: { type: [boolean, 'null'] }
        dimensions:
          type: array
          minItems: 3
          maxItems: 3
          items: { $ref: '#/components/schemas/CourseScoreDimension' }
    CourseScoreDimension:
      type: object
      additionalProperties: false
      required: [code, score, evidenceStatus, evidenceCoverage, components]
      properties:
        code: { type: string, enum: [USER_FEEDBACK, RELIABILITY, CLASS_QUALITY] }
        score: { type: number, minimum: 0 }
        evidenceStatus: { type: string }
        evidenceCoverage: { type: [string, 'null'] }
        components:
          type: array
          items: { $ref: '#/components/schemas/CourseScoreComponent' }
    CourseScoreComponent:
      type: object
      additionalProperties: false
      required: [code, score, pointsPerUnit, awarded, evidenceStatus]
      properties:
        code: { type: string }
        score: { type: number, minimum: 0 }
        pointsPerUnit: { type: [number, 'null'], minimum: 0 }
        awarded: { type: boolean }
        evidenceStatus: { type: string }
    Notification:
      type: object
      required: [sourceNotificationId, source, typeCode, title, body, relatedTaskCode, relatedTaskInstanceId, actionType, actionTarget, actionAvailable, expiresAt, expired, issuedAt, read, clicked]
      properties:
        sourceNotificationId: { type: string }
        source: { type: string, enum: [EXTERNAL, SYSTEM] }
        typeCode: { type: [string, 'null'] }
        title: { type: string }
        body: { type: string }
        relatedTaskCode: { type: [string, 'null'] }
        relatedTaskInstanceId: { type: [string, 'null'] }
        actionType:
          type: [string, 'null']
          enum: [TASK_DETAIL, MY_TIDE, TASKS, HELP, ACCOUNT, null]
        actionTarget: { type: [string, 'null'] }
        actionAvailable: { type: boolean }
        expiresAt: { type: [string, 'null'], format: date-time }
        expired: { type: boolean }
        issuedAt: { type: string, format: date-time }
        read: { type: boolean }
        clicked: { type: boolean }
...
    ApiError:
      type: object
      required: [code, message, requestId, retryable]
      properties:
        code: { type: string }
        message: { type: string }
        requestId: { type: string }
        retryable: { type: boolean }
        details: { type: object, additionalProperties: true }